Split up Elasticsearch module into multiple files
Was getting quite large
