+++ src/index.js
@@ -1,9 +1,12 @@
-require('newrelic');
+// require('newrelic');
 
 /* eslint-disable import/first */
 import app from './app';
 import { auditLogger } from './logger';
+import { initElasticsearchIntegration  } from './lib/elasticsearch';
 /* eslint-enable import/first */
+
+initElasticsearchIntegration();
 
 const port = process.env.PORT || 8080;
 const server = app.listen(port, () => {
