+++ src/lib/elasticsearch/queue.test.js
@@ -1,0 +1,8 @@
+import { createElasticsearchQueue } from "./queue";
+
+describe("Elasticsearch queue", () => {
+  test("gets created", () => {
+    const queue = createElasticsearchQueue();
+    expect(queue).not.toBeUndefined();
+  });
+});
